+++ R/validate.R
@@ -7,13 +7,17 @@ setValidity(
   Class = "AbundanceMatrix",
   method = function(object) {
     # Get data
+    n <- nrow(object)
     samples <- object@samples
     groups <- object@groups
 
-    cnd <- list(
-      catch_conditions(check_length(samples, nrow(object), strict = FALSE)),
-      catch_conditions(check_length(groups, nrow(object), strict = FALSE))
-    )
+    cnd <- list()
+    if (!is_empty(samples)) {
+      cnd <- c(cnd, validate(assert_length(samples, n)))
+    }
+    if (!is_empty(groups)) {
+      cnd <- c(cnd, validate(assert_length(groups, n)))
+    }
 
     # Return cnd, if any
     check_class(object, cnd)
@@ -24,21 +28,21 @@ setValidity(
 setValidity(
   Class = "IntegerMatrix",
   method = function(object) {
-    cnd <- list(catch_conditions(check_type(object, "integer")))
+    cnd <- list(validate(assert_type(object, "integer")))
     check_class(object, cnd)
   }
 )
 setValidity(
   Class = "NumericMatrix",
   method = function(object) {
-    cnd <- list(catch_conditions(check_type(object, "numeric")))
+    cnd <- list(validate(assert_type(object, "numeric")))
     check_class(object, cnd)
   }
 )
 setValidity(
   Class = "LogicalMatrix",
   method = function(object) {
-    cnd <- list(catch_conditions(check_type(object, "logical")))
+    cnd <- list(validate(assert_type(object, "logical")))
     check_class(object, cnd)
   }
 )
@@ -49,9 +53,7 @@ setValidity(
   Class = "CountMatrix",
   method = function(object) {
     cnd <- list(
-      catch_conditions(
-        check_numbers(object, "positive", strict = FALSE, na.rm = TRUE)
-      )
+      validate(assert_numeric(object, "positive", strict = FALSE, na.rm = TRUE))
     )
     check_class(object, cnd)
   }
@@ -64,12 +66,12 @@ setValidity(
     total <- object@total
 
     cnd <- list(
-      catch_conditions(check_symmetric(object)),
-      catch_conditions(
-        check_numbers(object, "positive", strict = FALSE, na.rm = TRUE)
-      ),
-      catch_conditions(check_scalar(total, "integer", strict = FALSE))
+      validate(assert_matrix(object, "symmetric")),
+      validate(assert_numeric(object, "positive", strict = FALSE, na.rm = TRUE))
     )
+    if (!is_empty(total)) {
+      cnd <- c(cnd, validate(assert_scalar(total, "integer")))
+    }
 
     # Return cnd, if any
     check_class(object, cnd)
@@ -82,15 +84,16 @@ setValidity(
   Class = "CompositionMatrix",
   method = function(object) {
     # Get data
+    n <- nrow(object)
     total <- object@total
 
     cnd <- list(
-      catch_conditions(
-        check_numbers(object, "positive", strict = FALSE, na.rm = TRUE)
-      ),
-      catch_conditions(check_numbers(total, "positive", strict = FALSE)),
-      catch_conditions(check_length(total, nrow(object), strict = FALSE))
+      validate(assert_numeric(object, "positive", strict = FALSE, na.rm = TRUE))
     )
+    if (!is_empty(total)) {
+      cnd <- c(cnd, validate(assert_numeric(total, "positive")))
+      cnd <- c(cnd, validate(assert_length(total, n)))
+    }
 
     # Return cnd, if any
     check_class(object, cnd)
@@ -102,17 +105,35 @@ setValidity(
 setValidity(
   Class = "StratigraphicMatrix",
   method = function(object) {
-    if (nrow(object) > 0) {
-      cnd <- list(
-        # Check data
-        catch_conditions(check_square(object)),
-        catch_conditions(check_dag(object))
-      )
-    } else {
-      cnd <- list()
+    cnd <- list()
+    if (!is_empty(object)) {
+      cnd <- c(cnd, validate(assert_matrix(object, "square")))
+      cnd <- c(cnd, validate(assert_dag(object)))
     }
 
     # Return cnd, if any
     check_class(object, cnd)
   }
 )
+
+# Diagnostic ===================================================================
+#' Class Diagnostic
+#'
+#' @param object An object to which error messages are related.
+#' @param conditions A [`list`] of condition messages.
+#' @return
+#'  Throw an error if `conditions` is of non-zero length, invisibly returns
+#'  `TRUE` if not.
+#' @author N. Frerebeau
+#' @keywords internal
+#' @noRd
+check_class <- function(object, conditions) {
+  cnd <- compact(is_empty, conditions)
+  if (has_length(cnd)) {
+    err <- sprintf("<%s> instance initialization:\n%s", class(object),
+                   paste0("* ", cnd, collapse = "\n"))
+    throw_error("arkhe_error_class", err, call = NULL)
+  }
+
+  invisible(TRUE)
+}